+++ src/pal/test/verify_uvm_attestation_and_endorsements.cpp
@@ -242,10 +242,13 @@ int main(int argc, char** argv)
 
         LOG_INFO_FMT("Verifying endorsements");
         const auto attestation_unverified =
-          ccf::pal::snp::AttestationReport::from_unverified(quote_info.quote);
+          ccf::pal::snp::parse_attestation_report_unverified(quote_info.quote);
         validate_endorsements(
           endorsements,
-          attestation_unverified.reported_tcb(),
+          ccf::pal::snp::TcbVersionRaw::from_span(
+            ccf::pal::snp::get_report_bytes(
+              attestation_unverified.get(),
+              tav_snp_attestation_report_reported_tcb)),
           quote_info.endorsements);
 
         LOG_INFO_FMT("Verifying quote");